--- src/controllers/token/erc20/_.validation.ts
@@ -1,13 +1,0 @@
-import { body, param } from 'express-validator';
-
-const validation = {
-    getERC20Token: [param('id').exists().isMongoId()],
-    postERC20Token: [
-        body('name').exists().isString(),
-        body('symbol').exists().isString(),
-        body('network').exists().isNumeric(),
-        body('totalSupply').exists().isString(),
-    ],
-};
-
-export default validation;
